split out misk-service module for misk's ServiceManager
