+++ misk-service/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=misk-service
+POM_NAME=misk-service
+POM_DESCRIPTION=A Misk module for service lifecycle management
+POM_PACKAGING=jar
