+++ settings.gradle
@@ -20,6 +20,7 @@ include ':misk-launchdarkly-core'
 include ':misk-metrics-digester'
 include ':misk-prometheus'
 include ':misk-redis'
+include ':misk-service'
 include ':misk-slack'
 include ':misk-testing'
 include ':misk-zipkin'
